# [Py] Optimize unpacking LOC-ID when logging Mils of logs