@gapisback 2024-04-04
@gregthelaw - Minor optimization fix in `l3_dump.py` -- to speed up dev-experience.

You can get to this whenever you have some precious spare cycles.