@gapisback 2024-04-07
Yes, @gregthelaw - I do agree. That thought did occur to me, to build a Python dictionary, but I figured, let's get this simple optimization in first, and then rework the logic depending on how much 'LOC-workload' we generate through more testing.

I prefer to build the test workload to create the situation of large #s of LOCs before building additional optimization. (Test-driven development.)